+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/download/manager/DownloadManagerRequestTask.java
@@ -6,8 +6,12 @@
 package com.microsoft.appcenter.distribute.download.manager;
 
 import android.app.DownloadManager;
+import android.database.Cursor;
 import android.net.Uri;
 import android.os.AsyncTask;
+import android.os.Handler;
+import android.os.Looper;
+
 import androidx.annotation.VisibleForTesting;
 
 import com.microsoft.appcenter.distribute.ReleaseDetails;
@@ -15,17 +19,23 @@ import com.microsoft.appcenter.utils.AppCenterLog;
 
 import static com.microsoft.appcenter.distribute.DistributeConstants.LOG_TAG;
 
+import java.util.concurrent.TimeoutException;
+
 /**
  * The download manager API triggers strict mode exception in UI thread.
  */
 class DownloadManagerRequestTask extends AsyncTask<Void, Void, Void> {
-
+    
+    private final int TIMEOUT_LIMIT = 10000;
     private final DownloadManagerReleaseDownloader mDownloader;
     private final String mTitle;
+    private Handler mHandler;
+    private Runnable handlerCallback;
 
     DownloadManagerRequestTask(DownloadManagerReleaseDownloader downloader, String title) {
         mDownloader = downloader;
         mTitle = title;
+        mHandler = new Handler(Looper.getMainLooper());
     }
 
     @Override
@@ -35,7 +45,7 @@ class DownloadManagerRequestTask extends AsyncTask<Void, Void, Void> {
         ReleaseDetails releaseDetails = mDownloader.getReleaseDetails();
         Uri downloadUrl = releaseDetails.getDownloadUrl();
         AppCenterLog.debug(LOG_TAG, "Start downloading new release from " + downloadUrl);
-        DownloadManager downloadManager = mDownloader.getDownloadManager();
+        final DownloadManager downloadManager = mDownloader.getDownloadManager();
         DownloadManager.Request request = createRequest(downloadUrl);
         request.setTitle(String.format(mTitle, releaseDetails.getShortVersion(), releaseDetails.getVersion()));
 
@@ -46,12 +56,28 @@ class DownloadManagerRequestTask extends AsyncTask<Void, Void, Void> {
         }
         long enqueueTime = System.currentTimeMillis();
         try {
-            long downloadId = downloadManager.enqueue(request);
+            final long downloadId = downloadManager.enqueue(request);
             if (!isCancelled()) {
                 mDownloader.onDownloadStarted(downloadId, enqueueTime);
+                handlerCallback = new Runnable() {
+
+                    @Override
+                    public void run() {
+                        DownloadManager.Query query = new DownloadManager.Query();
+                        query.setFilterByStatus(DownloadManager.STATUS_PENDING);
+                        Cursor c = downloadManager.query(query);
+                        if (c.moveToFirst()) {
+                            downloadManager.remove(downloadId);
+                            mDownloader.onDownloadError(new IllegalStateException("Failed to start downloading file due to timeout exception."));
+                        }
+                    }
+                };
+
+                /* Check that the file started to download. */
+                mHandler.postDelayed(handlerCallback, TIMEOUT_LIMIT);
             }
         } catch (IllegalArgumentException e) {
-            
+
             /*
              * In cases when Download Manager application is disabled,
              * IllegalArgumentException: Unknown URL content://downloads/my_download is thrown.
@@ -63,6 +89,9 @@ class DownloadManagerRequestTask extends AsyncTask<Void, Void, Void> {
 
     @VisibleForTesting
     DownloadManager.Request createRequest(Uri Uri) {
+        if (handlerCallback != null) {
+            mHandler.removeCallbacks(handlerCallback);
+        }
         return new DownloadManager.Request(Uri);
     }
 }