+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### App Center Distribute
 
+* **[Feature]** Remove the download manager task if the download doesn't start within 10 seconds.
 * **[Feature]** Replace install a new release from deprecated intent action [ACTION_INSTALL_PACKAGE](https://developer.android.com/reference/android/content/Intent#ACTION_INSTALL_PACKAGE) to `PackageInstaller` API.
 
 ___